services/ingest/pyproject.toml
[project]
name = "ai-knowledge-base-ingest"
version = "0.1.0"
description = "AI Knowledge Base Ingest Pipeline: Drive → S3 → Unstructured → OpenAI Vector Store"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.40.61",
    "google-api-python-client>=2.185.0",
    "google-auth>=2.41.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "openai>=2.6.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "tqdm>=4.67.1",
    "unstructured[pdf,docx,pptx,epub]>=0.18.15",
    "pypandoc>=1.13",
]

[project.optional-dependencies]
docling = [
    "docling>=2.0.0",
]
